Add files via upload
Adding a filter class for PySongFinder. It currently implements three filters: fold -- folds frequencies above the hearing range down to the selected low end of the range adding to current content; linear -- compresses all frequencies from the low end of the range to the limit for bird song (12kHz, supposedly) into equal-sized bins up to the hearing upper limit; and nonlinear -- same compression but with bins compressed at the upper end so that lower frequencies are unaffected.