Update OlaFFT.py
Fixed glitch caused by not having a continuous "signal" in the test main(). Resorted to generating a continuous "signal" and parsing out buffer loads. Still have an apparent phase shift at the end of the :signal." This may just be an artifact of FFT?

Commented out the zero of the trailing area of the buffer in irfft since that area is replaced in rfft.